remove list of devices from nav and add it to leftnav
This removes the redundant "welcome" mssage in leftnav and replaces it with a list of hyperlinks representing the download pages for each device. Only the "HOME" page in nav is left. This leaves a space for a redirect to your git, forums, etc, but i'll leave that up to you.

If you would like a good look @ what this looks like, you can see this on my site, buildbot.jdkoreclipse.net